Update refefences after renaming to VisuallyHiddenText
