+++ .storybook/DocsDecorator/HealthFactors/SingleDoc/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Text from '../../../../src/app/components/Text';
 import { Confirm, Close, ExternalLink, Help } from '../Icons';
 import styles from './index.styles';
-import VisuallyHiddenText from '../../../../src/app/components/psammead-visually-hidden-text';
+import VisuallyHiddenText from '../../../../src/app/components/VisuallyHiddenText';
 import idSanitiser from '../../../../src/app/lib/utilities/idSanitiser';
 
 const SingleDoc = ({
