+++ src/app/legacy/components/SkipLinkWrapper/index.jsx
@@ -4,7 +4,7 @@ import styled from '@emotion/styled';
 import detokenise from '#psammead/psammead-detokeniser/src';
 import { getSansBold } from '#psammead/psammead-styles/src/font-styles';
 import { GEL_BREVIER } from '#psammead/gel-foundations/src/typography';
-import { visuallyHiddenTextStyle } from '../../../components/psammead-visually-hidden-text';
+import { visuallyHiddenTextStyle } from '../../../components/VisuallyHiddenText';
 
 const BORDER_WIDTH = '0.125rem';
 const GEL_SPACING_PLUS_HALF = `0.75rem`;
