+++ src/app/legacy/containers/Copyright/index.jsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import { string } from 'prop-types';
 import Copyright from '#psammead/psammead-copyright/src';
-import VisuallyHiddenText from '../../../components/psammead-visually-hidden-text';
+import VisuallyHiddenText from '../../../components/VisuallyHiddenText';
 import { ServiceContext } from '../../../contexts/ServiceContext';
 
 const CopyrightContainer = ({ children }) => {
