+++ src/app/legacy/psammead/psammead-caption/src/index.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import VisuallyHiddenText from '../../../../components/psammead-visually-hidden-text';
+import VisuallyHiddenText from '../../../../components/VisuallyHiddenText';
 import { render } from '../../../../components/react-testing-library-with-providers';
 import latin from '../../../../components/ThemeProvider/fontScripts/latin';
 import arabic from '../../../../components/ThemeProvider/fontScripts/arabic';
