+++ src/app/legacy/psammead/psammead-navigation/src/index.jsx
@@ -14,7 +14,7 @@ import {
 import { getPica } from '#psammead/gel-foundations/src/typography';
 import { scriptPropType } from '#psammead/gel-foundations/src/prop-types';
 import { getSansRegular } from '#psammead/psammead-styles/src/font-styles';
-import VisuallyHiddenText from '../../../../components/psammead-visually-hidden-text';
+import VisuallyHiddenText from '../../../../components/VisuallyHiddenText';
 import { NAV_BAR_TOP_BOTTOM_SPACING } from './DropdownNavigation';
 import { focusIndicatorThickness } from '../../../../components/ThemeProvider/focusIndicator';
