+++ src/app/legacy/psammead/psammead-social-embed/src/utilities/index.js
@@ -60,7 +60,7 @@ export const getCaptionText = ({ pageType, caption, provider }) => {
 };
 
 /**
- * These styles are lifted from #psammead/psammead-visually-hidden-text, which we
+ * These styles are lifted from #psammead/VisuallyHiddenText, which we
  * cannot use directly as we require control over _when_ they are applied.
  */
 export const visuallyHiddenStyle = `
