+++ src/app/legacy/psammead/psammead-story-promo/src/IndexAlsos/index.jsx
@@ -9,7 +9,7 @@ import {
 } from '#psammead/gel-foundations/src/spacings';
 import { getBrevier } from '#psammead/gel-foundations/src/typography';
 import { getSerifMedium } from '#psammead/psammead-styles/src/font-styles';
-import VisuallyHiddenText from '../../../../../components/psammead-visually-hidden-text';
+import VisuallyHiddenText from '../../../../../components/VisuallyHiddenText';
 
 const StyledIndexAlsos = styled.div`
   position: relative;
